# Faster paths for nested parsers