@scotttrinh 2022-03-21
I wonder why `number` and `string` got slower in the benchmarks? Maybe the overhead of instantiating the `ParseInputLazyPath` class?