@scotttrinh 2022-03-21
Yeah, I came away with similar results. I was hoping the benchmarking library itself doing a bunch of sampling runs would help alleviate this somewhat, but 🤷 I know it's common to disregard "microbenchmarks", but I have a lot of use cases that use single primitive Zod schemas, so I still personally care about the primitive performance apart from how they are composed into something like `realworld`.

As an aside: my dream here is to run benchmarks in CI and show diffs rather than the raw data. That would at least alleviate some of the issues with having other processes interfering with results.

Thanks for all of your great performance work @tmcw!